Add setup script

+++ start.sh
@@ -6,12 +6,14 @@
 : ${PASSWORD:?"Error: PASSWORD environment variable is not set."}
 
 : ${ROOT_PASSWORD:?"Error: ROOT_PASSWORD environment variable is not set."}
-: ${CS_PASSWORD:?"Error: CS_PASSWORD environment variable is not set."}
+: ${CS_PASSWORD:=""}
 
 : ${AUTHORIZED_KEYS:=""}
 
 : ${GIT_USERNAME:=""}
 : ${GIT_EMAIL:=""}
+
+: ${SETUP_SCRIPT:="~/.devbox/setup.sh"}
 
 # Set the root password
 echo "root:${ROOT_PASSWORD}" | chpasswd
@@ -40,7 +42,9 @@ if [ -n "${GIT_EMAIL}" ]; then
 fi
 
 # Start code-server
-su ${USERNAME} -c 'PASSWORD=${CS_PASSWORD} nohup code-server --bind-addr=0.0.0.0:8080 --app-name=devbox --auth=password &'
+if [ -n "${CS_PASSWORD}" ]; then
+    su ${USERNAME} -c 'PASSWORD=${CS_PASSWORD} nohup code-server --bind-addr=0.0.0.0:8080 --app-name=devbox --auth=password &'
+fi
 
 # Create .ssh folder
 mkdir -p /home/${USERNAME}/.ssh
@@ -48,11 +52,18 @@ chmod 700 /home/${USERNAME}/.ssh
 chown -R ${USERNAME}:${USERNAME} /home/${USERNAME}/.ssh
 
 # Set the authorized keys from the AUTHORIZED_KEYS environment variable (if provided)
-if [ -n "${AUTHORIZED_KEYS}" ]; then
+if [ -n "${AUTHORIZED_KEYS}" ] && [ ! -f "/home/${USERNAME}/.ssh/authorized_keys" ]; then
     echo "${AUTHORIZED_KEYS}" > /home/${USERNAME}/.ssh/authorized_keys
     chown -R ${USERNAME}:${USERNAME} /home/${USERNAME}/.ssh
     chmod 600 /home/${USERNAME}/.ssh/authorized_keys
     echo "Authorized keys set for user ${USERNAME}"
+fi
+
+# Call the setup script
+if [ -n "${SETUP_SCRIPT}" ]; then
+    chmod +x "${SETUP_SCRIPT}"
+    echo "Calling setup script"
+    "${SETUP_SCRIPT}"
 fi
 
 # Start the SSH server
